+++ domain/exceptions.py
@@ -1,0 +1,10 @@
+class FileUploadError(Exception):
+    pass
+
+
+class InvalidFileTypeError(FileUploadError):
+    pass
+
+
+class FileSaveError(FileUploadError):
+    pass